Sweep: Remove unused transformThroughLine function from production
Moved to test file where it is still used as a helper.
